[CMake] Split add_swift_library in two
This patch splits add_swift_library into two functions, add_swift_library and add_swift_target_library. The later is very complicated and its complication was bleeding into the code needed for the former.

Splitting the two up dramatically simplifies the build of the Swift compiler and limits the complexity to just the runtime build.